code cleanup and reworked the _read_mailbox
